All featuretests declaration to move in the same file.
